// __generated__/StoreQuery.graphql.ts
export type StoreQuery$variables = {
  id: string;
};
export type StoreQuery$data = {
  readonly storesCollection: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly address: string | null | undefined;
        readonly delivery_time_max: number | null | undefined;
        readonly delivery_time_min: number | null | undefined;
        readonly id: string;
        readonly image: string | null | undefined;
        readonly is_active: boolean | null | undefined;
        readonly is_available: boolean | null | undefined;
        readonly latitude: string | null | undefined;
        readonly longitude: string | null | undefined;
        readonly minimum_order: string | null | undefined;
        readonly name: string;
        readonly rating: string | null | undefined;
        readonly review_count: number | null | undefined;
        readonly slug: string | null | undefined;
        readonly store_categories_id: string | null | undefined;
      };
    }>;
  } | null | undefined;
};
export type StoreQuery = {
  response: StoreQuery$data;
  variables: StoreQuery$variables;
};